+++ src/protocols/writing.rs
@@ -65,19 +65,21 @@ where
                     trace!(parent: node.span(), "spawned a task for writing messages to {}", addr);
                     tx_writer.send(()).unwrap(); // safe; the channel was just opened
 
-                    while let Some(msg) = outbound_message_receiver.recv().await {
-                        let msg = msg.downcast::<Self::Message>().unwrap();
+                    while let Some(wrapped_msg) = outbound_message_receiver.recv().await {
+                        let msg = wrapped_msg.msg.downcast::<Self::Message>().unwrap();
 
                         match writer_clone
                             .write_to_stream(*msg, addr, &mut buffer, &mut writer)
                             .await
                         {
                             Ok(len) => {
+                                let _ = wrapped_msg.delivery_notification.send(true);
                                 node.known_peers().register_sent_message(addr, len);
                                 node.stats().register_sent_message(len);
                                 trace!(parent: node.span(), "sent {}B to {}", len, addr);
                             }
                             Err(e) => {
+                                let _ = wrapped_msg.delivery_notification.send(false);
                                 node.known_peers().register_failure(addr);
                                 error!(parent: node.span(), "couldn't send a message to {}: {}", addr, e);
                                 if node.config().fatal_io_errors.contains(&e.kind()) {
@@ -141,7 +143,8 @@ where
     ) -> io::Result<()>;
 
     /// Sends the provided message to the specified `SocketAddr`. Returns as soon as the message is queued to
-    /// be sent, without waiting for the actual delivery.
+    /// be sent, without waiting for the actual delivery; instead, the caller is provided with a [`oneshot::Receiver`]
+    /// which can be used to determine when and whether the message has been delivered.
     ///
     /// # Errors
     ///
@@ -152,16 +155,17 @@ where
         &self,
         addr: SocketAddr,
         message: Self::Message,
-    ) -> io::Result<oneshot::Receiver<()>> {
+    ) -> io::Result<oneshot::Receiver<bool>> {
         // access the protocol handler
         if let Some(handler) = self.node().protocols.writing_handler.get() {
             // find the message sender for the given address
             if let Some(sender) = handler.senders.read().get(&addr).cloned() {
-                sender.try_send(Box::new(message)).map_err(|e| {
+                let (msg, delivery) = WrappedMessage::new(Box::new(message));
+                sender.try_send(msg).map_err(|e| {
                     error!(parent: self.node().span(), "can't send a message to {}: {}", addr, e);
                     self.node().stats().register_failure();
                     io::ErrorKind::Other.into()
-                })
+                }).map(|_| delivery)
             } else {
                 Err(io::ErrorKind::NotConnected.into())
             }
@@ -184,7 +188,8 @@ where
         if let Some(handler) = self.node().protocols.writing_handler.get() {
             let senders = handler.senders.read().clone();
             for (addr, message_sender) in senders {
-                let _ = message_sender.try_send(Box::new(message.clone())).map_err(|e| {
+                let (msg, _delivery) = WrappedMessage::new(Box::new(message.clone()));
+                let _ = message_sender.try_send(msg).map_err(|e| {
                     error!(parent: self.node().span(), "can't send a message to {}: {}", addr, e);
                     self.node().stats().register_failure();
                 });
@@ -197,10 +202,28 @@ where
     }
 }
 
+/// Used to queue messages for delivery.
+pub(crate) struct WrappedMessage {
+    msg: Box<dyn Any + Send>,
+    delivery_notification: oneshot::Sender<bool>,
+}
+
+impl WrappedMessage {
+    fn new(msg: Box<dyn Any + Send>) -> (Self, oneshot::Receiver<bool>) {
+        let (tx, rx) = oneshot::channel();
+        let wrapped_msg = Self {
+            msg,
+            delivery_notification: tx,
+        };
+
+        (wrapped_msg, rx)
+    }
+}
+
 /// The handler object dedicated to the `Writing` protocol.
 pub struct WritingHandler {
     handler: mpsc::Sender<ReturnableConnection>,
-    pub(crate) senders: RwLock<HashMap<SocketAddr, mpsc::Sender<Box<dyn Any + Send>>>>,
+    pub(crate) senders: RwLock<HashMap<SocketAddr, mpsc::Sender<WrappedMessage>>>,
 }
 
 impl WritingHandler {